Separada la logida de JavaScript en un hook personalizado: useMainArticleLogic.js
